+++ src/com/pacman/systemelements/Dimension.java
@@ -8,19 +8,19 @@ public final class Dimension {
     /**
      * Largura do {@link SceneElement}.
      */
-    private final float width;
+    private final double width;
 
     /**
      * Altura do {@link SceneElement}.
      */
-    private final float height;
+    private final double height;
 
     /**
      * Construtor padrão. Os valores de altura e largura serão ambos inicializados com 0.0f.
      */
     public Dimension() {
-        this.width = 0.0f;
-        this.height = 0.0f;
+        this.width = 0.0;
+        this.height = 0.0;
     }
 
     /**
@@ -29,7 +29,7 @@ public final class Dimension {
      * @param height altura do {@link SceneElement}.
      * @throws IllegalArgumentException caso algum dos parâmetros passados seja menor que zero.
      */
-    public Dimension(float width, float height) throws IllegalArgumentException {
+    public Dimension(double width, double height) throws IllegalArgumentException {
         if (width < 0) throw new IllegalArgumentException("Width can not be a negative number");
         if (height < 0) throw new IllegalArgumentException("Height can not be a negative number");
         this.width = width;
@@ -40,7 +40,7 @@ public final class Dimension {
      * Retorna a largura do {@link SceneElement}.
      * @return a largura do {@link SceneElement}.
      */
-    public float getWidth() {
+    public double getWidth() {
         return width;
     }
 
@@ -48,7 +48,7 @@ public final class Dimension {
      * Retorna a altura do {@link SceneElement}.
      * @return a altura do {@link SceneElement}.
      */
-    public float getHeight() {
+    public double getHeight() {
         return height;
     }
